requirements.txt

# run this line separately to install pytorch
# gpu version
# pip3 install torch==1.8.1+cu111 torchvision==0.9.1+cu111 torchaudio==0.8.1 -f https://download.pytorch.org/whl/torch_stable.html
# pip install torch==1.13.1+cu117 torchvision==0.14.1+cu117 torchaudio==0.13.1 --extra-index-url https://download.pytorch.org/whl/cu117
#
# cpu version
# pip install torch==1.8.0+cpu torchvision==0.9.0+cpu torchaudio==0.8.0 -f https://download.pytorch.org/whl/torch_stable.html
######

# torch==1.4.0
transformers==4.6.1 #old 2.9.0
pytorch_lightning==1.6.4 #0.7.5 #1.3.2 
sentencepiece==0.1.95
wordfreq==2.5.1
python-magic==0.4.24 # detect file encoding
pandas==1.5.3
tqdm==4.64.0
optuna==3.1.0
optuna-dashboard==0.8.1
wandb
pyhyphen==3.0.1
# nltk>=3.6.5
ftfy
scikit-learn 
scipy 
matplotlib
diskcache
